move pDAO proposal actions from events to functions
