+++ rocketwatch/plugins/events/events.json
@@ -301,31 +301,6 @@
                 {
                     "event_name": "ActionRequestLeave",
                     "name": "sdao_member_request_leave_event"
-                },
-                {
-                    "event_name": "ActionKick",
-                    "name": "sdao_member_kick_event"
-                }
-            ]
-        },
-        {
-            "contract_name": "rocketDAOProtocolProposals",
-            "events": [
-                {
-                    "event_name": "ProposalSettingUint",
-                    "name": "pdao_setting_event"
-                },
-                {
-                    "event_name": "ProposalSettingBool",
-                    "name": "pdao_setting_event"
-                },
-                {
-                    "event_name": "ProposalSettingAddress",
-                    "name": "pdao_setting_event"
-                },
-                {
-                    "event_name": "ProposalSettingRewardsClaimers",
-                    "name": "pdao_claimer_event"
                 }
             ]
         },
@@ -343,10 +318,6 @@
                 {
                     "event_name": "ProposalVoteOverridden",
                     "name": "pdao_proposal_vote_override_event"
-                },
-                {
-                    "event_name": "ProposalExecuted",
-                    "name": "pdao_proposal_execute_event"
                 },
                 {
                     "event_name": "ProposalFinalised",
@@ -379,20 +350,8 @@
             "contract_name": "rocketClaimDAO",
             "events": [
                 {
-                    "event_name": "RPLTokensSentByDAOProtocol",
-                    "name": "pdao_spend_treasury_event"
-                },
-                {
                     "event_name": "RPLTreasuryContractClaimed",
                     "name": "pdao_spend_treasury_recurring_claim_event"
-                },
-                {
-                    "event_name": "RPLTreasuryContractCreated",
-                    "name": "pdao_spend_treasury_recurring_new_event"
-                },
-                {
-                    "event_name": "RPLTreasuryContractUpdated",
-                    "name": "pdao_spend_treasury_recurring_update_event"
                 }
             ]
         }
